Reintroduce CredentialIssuerMetadataJsonParser.
